chore: remove trading-form-group
