--- libs/ui-toolkit/src/components/trading-form-group/index.tsx
@@ -1,1 +1,0 @@
-export * from './form-group';
