+++ libs/withdraws/src/lib/withdraw-form.tsx
@@ -10,7 +10,7 @@ import {
 } from '@vegaprotocol/utils';
 import { useLocalStorage } from '@vegaprotocol/react-helpers';
 import {
-  TradingFormGroup,
+  FormGroup,
   TradingInput,
   TradingInputError,
   Notification,
@@ -195,7 +195,7 @@ export const WithdrawForm = ({
         noValidate={true}
         data-testid="withdraw-form"
       >
-        <TradingFormGroup
+        <FormGroup
           label={t('To {{chainName}} address', {
             chainName: selectedAsset?.chainId
               ? getChainName(selectedAsset.chainId)
@@ -278,9 +278,9 @@ export const WithdrawForm = ({
               {errors.to.message}
             </TradingInputError>
           )}
-        </TradingFormGroup>
+        </FormGroup>
 
-        <TradingFormGroup label={t('Asset')} labelFor="asset">
+        <FormGroup label={t('Asset')} labelFor="asset">
           <Controller
             control={control}
             name="asset"
@@ -337,7 +337,7 @@ export const WithdrawForm = ({
               {errors.asset.message}
             </TradingInputError>
           )}
-        </TradingFormGroup>
+        </FormGroup>
 
         {selectedAsset && (
           <div className="mb-4">
@@ -351,7 +351,7 @@ export const WithdrawForm = ({
             />
           </div>
         )}
-        <TradingFormGroup label={t('Amount')} labelFor="amount">
+        <FormGroup label={t('Amount')} labelFor="amount">
           <TradingInput
             data-testid="amount-input"
             type="number"
@@ -397,7 +397,7 @@ export const WithdrawForm = ({
               />
             </div>
           )}
-        </TradingFormGroup>
+        </FormGroup>
 
         <Button data-testid="submit-withdrawal" type="submit" fill={true}>
           {t('Release funds')}
